Added a new map called weather.html
This map uses the openweathermap API to add precipitation, wind and temperature layer on the top of openstreetmap layer.
Plus instead of display a standard marker it displays at icon based on the weather in that location.
Unfortunately the free API for owt doesn't let you use the new and shiny weather map 2 so then we can have only have current forcast.