Adding simple circle script to update website to a s3 bucket.
The circleci script has 1 job called build which check out the github project, put all the files need for the website in the folder website and then call the aws s3 sync, to "sync" the files with what is already in the bucker :)
In order to make it work with you repo you need to setup a circleci account and circleci user in AWS and then grant him a policy to update files to s3. (I can show you Sunday :))
At the moment the deployment happen to my aws:
https://s3.amazonaws.com/com.multipitch/builds/index.html
We can think in the future to have to location one for test and the other for prod :)

ps: some of the path are broken because are decleated with src="/" but I guess it was done on purpuse becauseyou have another bucket with those images?