Added a simple way to generate the files in data/
I am using a library to read the amazing excel file and generate the data files as they were.
In order to run you have to type: npm install to install the npm dependencies (the excel reader, also this need to be done only once)
then you have to run npm start or node convertXlsxToJson.js to generate the files.
when we have CD we are going to add that step in Circle CI.