TEST-TO-BE-DISCUSSED - added js/load-data to download the data files from a Google spreesheet.
The multi-pitch-data.xlsx was updated to google sheet: https://docs.google.com/spreadsheets/d/1uWQR5-77i5gL4tp52xcjH4_5qJMh6S2LHEXQJUqpzsc/edit#gid=886877247
Bare in mind is public so you may want to take it out.
Google Sheet has a REST api that you can call to get the sheets rows in JSON then I attached the result to the window object of the broweser,
so existent functionality is still preserved. The only bad things is that it takes longer to download... but I am sure we can work on that with cache or some other clever things.
The good things of this is that it doesn not require a website deployment to add new climbs :)