fixed img path to be relative and not absolute
I am not sure if I am going to brake aws but the address of the images should be relative to the project and not absolute, so then it can be easyly tested locally.
mainly replace /img with ../img or ./img.
Also I close a div in the section that was braking old version of chrome