create a component folder where put all the html for craeting a climb card
I create a new folder called components where we can put all the html templating.
So then the main can be a little more dumb and require templates from there.
Also this is the first step to have a html for every climb. the idea is that they are going to reuse the html from that folder.