Created a node file generateCardClimb.js to generate card html and put them in website/climbs.
the folder website/climbs like the website/data is not in git because it is autogenerated.
Also in order to do this changes I had to use the rootProject because the imgs are going to have different roots if they are in main or if they are in climbs.
the name is generated with this function:
"".concat(climbAndHtml.climb.cliff, '-', climbAndHtml.climb.routeName, '.html')
      .toLowerCase()
      .replace(/ /g, "_");
you can generate the html with this command: npm run generate-climbs

At the moment the generated html doesn't have css and is not linked to the index.. this is because I have no idea what to do next :)