Need a longer explanation and I need to go work...
basically all the js was moved to a new folder called src.
then webpack comes in and boundle all the src and css in 2 files and move then to website/dist.

DOing that we have around 50% smaller boundle to be loaded and there are many more things that we can do with webpack.