Comparison type == "temperature" may cause unexpected type coercion less
It is considered a good practice to use the type-safe equality operators === and !== instead of their regular counterparts == and !=.
type-safe equality
A strict comparison (e.g., ===) is only true if the operands are of the same type and the contents match. The more commonly-used abstract comparison (e.g. ==) converts the operands to the same type before making the comparison.
you do not want to covert to the same type :)