Done some formatting code and added auth-stuff.js to be loaded in the main page.
Auth-stuff on windows load is going to check if there is a auth token in the local storage or in the url.
if the token is in the url or in the local storage, it checks if it is valid and if it is valid it will load the user preferences and print in the console.
In the case the token is not present in both places,it is going to print a message with the url to signin.